Abstract away BFS
